+++ packages/sammi-next/src/node/constants.ts
@@ -1,21 +1,25 @@
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
-import colors from 'picocolors'
 
 export { version as VERSION } from '../../package.json';
 
 export const DEFAULT_CONFIG_EXTENSIONS: string[] = [
-    '.mjs',
-    '.js',
     '.mts',
     '.ts',
+    '.cts',
+    '.mjs',
+    '.js',
+    '.cjs',
+    '.json',
 ]
 
 export const GLOBAL_NAME = "SAMMIExtensions";
 
-export const BUILD_PREFIX = colors.blue("[sammi-next]");
-export const GREEN_CHECK = colors.green('✔');
-export const RED_X = colors.red('✗');
+export const LOGGER_PREFIX = '[sammi-next]';
+export const INFO_PREFIX = 'ℹ';
+export const SUCCESS_PREFIX = '✔';
+export const WARN_PREFIX = '⚠';
+export const ERROR_PREFIX = '✗';
 
 function findPackageDir() {
     let initPath = fileURLToPath(import.meta.url);