+++ packages/sammi-next/src/shared/build-types.ts
@@ -1,0 +1,6 @@
+export enum BuildMode {
+    dev,
+    production,
+}
+export const BuildModes = Object.keys(BuildMode).filter(key => isNaN(Number(key)));
+export type BuildModeKey = keyof typeof BuildMode;